BONUS: Refactor the common test functions as a shared module
This helps writing the tests in the next commits.
